Sound support

+++ apple2/src/misc.rs
@@ -2,12 +2,12 @@ use std::fs::File;
 use std::io::Write;
 use crate::constants::CYCLES;
 
-pub(crate) fn increase_cycles(v: u128) {
+pub(crate) fn increase_cycles(v: u64) {
     let c = CYCLES.read().unwrap().wrapping_add(v);
     *CYCLES.write().unwrap() = c;
 }
 
-pub(crate) fn get_cycles() -> u128 {
+pub(crate) fn get_cycles() -> u64 {
     *CYCLES.read().unwrap()
 }
